requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alpha_vantage==3.0.0
attrs==24.3.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
chex==0.1.88
click==8.1.7
contourpy==1.3.1
cycler==0.12.1
dm-tree==0.1.9
docstring_parser==0.16
etils==1.12.0
Flask==3.1.0
flax==0.10.2
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2025.2.0
googleapis-common-protos==1.66.0
humanize==4.11.0
idna==3.10
immutabledict==4.2.1
importlib_resources==6.5.2
itsdangerous==2.2.0
jax==0.5.0
jaxlib==0.5.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.8
LinearRegression==0.0.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
ml_dtypes==0.5.1
msgpack==1.1.0
multidict==6.1.0
nest-asyncio==1.6.0
numpy==2.2.1
opt_einsum==3.4.0
optax==0.2.4
orbax-checkpoint==0.11.4
packaging==24.2
pandas==2.2.3
pillow==11.1.0
promise==2.3
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
pyarrow==19.0.0
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
setuptools==75.8.0
simple-parsing==0.1.7
simplejson==3.19.3
six==1.17.0
tensorflow-datasets==4.9.7
tensorflow-metadata==1.16.1
tensorstore==0.1.71
termcolor==2.5.0
threadpoolctl==3.5.0
toml==0.10.2
toolz==1.0.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
